% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareEnvironment.R
\name{CheckAndUpdatePackage}
\alias{CheckAndUpdatePackage}
\title{Check and Update Package from GitHub}
\usage{
CheckAndUpdatePackage(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. Whether to suppress messages}
}
\description{
Internal function to check for package updates and install if available
}
\keyword{internal}
